back/requirements.txt
Django==3.2.8
djangorestframework==3.12.4
djongo==1.3.6
django-cors-headers==3.8.0
pymongo==3.12.3
keras==2.6.0
tensorflow

